rename TS file
